# signing-service-challenge-go/docs/swagger.yaml
basePath: /api/v0
definitions:
  api.CreateDeviceResponse:
    properties:
      algorithm:
        type: string
      id:
        type: string
      label:
        type: string
      privateKey:
        type: string
      publicKey:
        type: string
    type: object
  api.ErrorResponse:
    properties:
      errors:
        items:
          type: string
        type: array
    type: object
  api.GetAllDevicesResponse:
    properties:
      devices:
        items:
          $ref: '#/definitions/api.GetDeviceResponse'
        type: array
      total:
        type: integer
    type: object
  api.GetDeviceResponse:
    properties:
      algorithm:
        type: string
      id:
        type: string
      label:
        type: string
      lastSignature:
        type: string
      privateKey:
        type: string
      publicKey:
        type: string
      signatureCounter:
        type: integer
    type: object
  api.HealthResponse:
    properties:
      status:
        type: string
      version:
        type: string
    type: object
  api.SignaturedDataResponse:
    properties:
      signature:
        items:
          type: integer
        type: array
      signed_data:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for managing signature devices and signing transactions.
  title: Signing Service API
  version: "1.0"
paths:
  /{deviceId}:
    get:
      description: Retrieves a device by its ID and returns its details.
      parameters:
      - description: Device ID
        in: path
        name: deviceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Device successfully retrieved
          schema:
            $ref: '#/definitions/api.GetDeviceResponse'
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Device not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get a device
      tags:
      - Devices
  /all:
    get:
      description: Retrieves all the devices and its details.
      produces:
      - application/json
      responses:
        "200":
          description: Devices successfully retrieved
          schema:
            $ref: '#/definitions/api.GetAllDevicesResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get all the devices
      tags:
      - Devices
  /health:
    get:
      consumes:
      - application/json
      description: Evaluates the health of the service and returns a standardized
        response.
      produces:
      - application/json
      responses:
        "200":
          description: Service is healthy
          schema:
            $ref: '#/definitions/api.HealthResponse'
        "405":
          description: Method not allowed
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Check the health of the service
      tags:
      - Health
  /new-device:
    post:
      description: Creates a new signature device with the specified parameters
      parameters:
      - description: Algorithm (ECC or RSA)
        in: query
        name: algorithm
        required: true
        type: string
      - description: Label for the device
        in: query
        name: label
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.CreateDeviceResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create a new signature device
      tags:
      - Devices
  /sign/{deviceId}:
    post:
      description: Signs a transaction using the specified device ID and data payload.
      parameters:
      - description: Device ID
        in: path
        name: deviceId
        required: true
        type: string
      - description: Data to be signed
        in: body
        name: data
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: Signature successfully generated
          schema:
            $ref: '#/definitions/api.SignaturedDataResponse'
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Device not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Sign a transaction
      tags:
      - Devices
swagger: "2.0"